+++ 03-components/30-removable-emails/MarkedEmailsApp.js
@@ -69,7 +69,7 @@ export default defineComponent({
       <UiFormGroup>
         <UiInput v-model.trim="query" type="search" placeholder="Поиск" aria-label="Поиск" small />
       </UiFormGroup>
-      <EmailList :emails="markedEmails" @removeEmail="removeEmailByIndex($event)" />
+      <EmailList :emails="markedEmails" @removeEmail="removeEmailByIndex" />
     </div>
   `,
 })
