+++ 03-components/50-weather-components/WeatherCard.js
@@ -1,7 +1,19 @@
-import { defineComponent } from 'vue'
+import {defineComponent, toRef, computed} from 'vue'
 import { WeatherConditionIcons } from './weather.service.ts'
 import WeatherAlert from './WeatherAlert.js'
 import WeatherDetailsItem from './WeatherDetailsItem.js'
+
+function isNight(dt, sunrise, sunset) {
+  return dt < sunrise || dt > sunset
+}
+
+function convertKelvinToCelsius(kelvin) {
+  return (kelvin - 273.15).toFixed(1)
+}
+
+function converthPatommHg(hPa) {
+  return (hPa * 0.75).toFixed(0)
+}
 
 export default defineComponent({
   name: 'WeatherCard',
@@ -19,18 +31,29 @@ export default defineComponent({
   },
 
   setup(props) {
-    const { alert, current, geographic_name } = props.weatherItem
+    const weatherItem = toRef(() => props.weatherItem)
+
+    const { alert, geographic_name, current } = weatherItem.value
+
+    const currentWeatherItem = computed(() => {
+      return {
+        ...current,
+        temp: convertKelvinToCelsius(current.temp),
+        pressure: converthPatommHg(current.pressure),
+        isNight: isNight(current.dt, current.sunrise, current.sunset),
+      }
+    })
 
     return {
       WeatherConditionIcons,
       alert,
-      current,
+      currentWeatherItem,
       geographic_name,
     }
   },
 
   template: `
-    <li class="weather-card" :class="{'weather-card--night': current.isNight }">
+    <li class="weather-card" :class="{'weather-card--night': currentWeatherItem.isNight }">
       <WeatherAlert v-if="alert">
         {{ alert.sender_name }}: {{ alert.description }}
       </WeatherAlert>
@@ -39,19 +62,19 @@ export default defineComponent({
           {{ geographic_name }}
         </h2>
         <div class="weather-card__time">
-          {{ current.dt }}
+          {{ currentWeatherItem.dt }}
         </div>
       </div>
       <div class="weather-conditions">
-        <div class="weather-conditions__icon" :title="current.weather.description">{{ WeatherConditionIcons[current.weather.id] }}</div>
-        <div class="weather-conditions__temp">{{ current.temp }} °C</div>
+        <div class="weather-conditions__icon" :title="currentWeatherItem.weather.description">{{ WeatherConditionIcons[currentWeatherItem.weather.id] }}</div>
+        <div class="weather-conditions__temp">{{ currentWeatherItem.temp }} °C</div>
       </div>
 
       <div class="weather-details">
-        <WeatherDetailsItem title="Давление, мм рт. ст." :value="current.pressure" />
-        <WeatherDetailsItem title="Влажность, %" :value="current.humidity" />
-        <WeatherDetailsItem title="Облачность, %" :value="current.clouds" />
-        <WeatherDetailsItem title="Ветер, м/с" :value="current.wind_speed" />
+        <WeatherDetailsItem title="Давление, мм рт. ст." :value="currentWeatherItem.pressure" />
+        <WeatherDetailsItem title="Влажность, %" :value="currentWeatherItem.humidity" />
+        <WeatherDetailsItem title="Облачность, %" :value="currentWeatherItem.clouds" />
+        <WeatherDetailsItem title="Ветер, м/с" :value="currentWeatherItem.wind_speed" />
       </div>
     </li>
   `,
